Remove Static Pendable
